+++ crates/core/src/message/mod.rs
@@ -32,7 +32,6 @@ pub use types::*;
 pub mod handlers;
 pub use handlers::storage::ChordStorageInterface;
 pub use handlers::storage::ChordStorageInterfaceCacheChecker;
-pub use handlers::subring::SubringInterface;
 pub use handlers::HandleMsg;
 pub use handlers::MessageHandler;