+++ crates/snark/src/prelude.rs
@@ -1,6 +1,0 @@
-//! reimport libs here
-pub use bellman_ce as bellman;
-pub use bellpepper_core as bellpepper;
-pub use ff;
-pub use nova_snark as nova;
-pub use pasta_curves as curves;