+++ crates/snark/src/tests/native/mod.rs
@@ -1,3 +1,0 @@
-mod test_snark;
-mod test_wat;
-mod test_witness;